+++ aiosonic/__init__.py
@@ -1,4 +1,4 @@
-from aiosonic.base_client import BaseClient
+from aiosonic.base_client import AioSonicBaseClient, BaseClient
 from aiosonic.client import *
 from aiosonic.sse_client import *
 from aiosonic.web_socket_client import *